+++ src/vm/token_d.cpp
@@ -1,8 +1,9 @@
 
+#include <stdint.h>
 #include <esteh/vm/opcode.hpp>
 #include <esteh/vm/code_parser.hpp>
 
-int code_parser::token_d(char *token) {
+uint16_t code_parser::token_d(char *token) {
 
 
 	#define CP(val) (!strcmp(token, val))
@@ -10,6 +11,5 @@ int code_parser::token_d(char *token) {
 	if (CP("p") || CP("print") || CP("echo")) {
 		return TD_PRINT;
 	}
-
 	return T_UNKNOWN;
 }
